refactor: domain 패치키, view 패키지를 구성하여 비즈니스 로직과 UI 관련 객체를 분리
추가적인 dirty code & line 수정
