+++ src/main/java/racing/RacingApplication.java
@@ -1,5 +1,10 @@
 package racing;
 
+import racing.domain.Cars;
+import racing.view.InputView;
+import racing.domain.Racing;
+import racing.domain.condition.RandomCondition;
+
 public class RacingApplication {
     public static void main(String[] args) {
         Cars cars = new Cars(InputView.carNamesMessage(), new RandomCondition());
