+++ src/main/java/racing/domain/condition/MovingCondition.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain.condition;
 
 public interface MovingCondition {
     int getCondition();
