+++ src/main/java/racing/view/ResultView.java
@@ -1,4 +1,8 @@
-package racing;
+package racing.view;
+
+import racing.domain.Car;
+import racing.domain.Cars;
+import racing.domain.RaceWinner;
 
 import java.util.Iterator;
 
@@ -10,16 +14,16 @@ public class ResultView {
     private static final String RACING_END_MESSAGE = "가 최종 우승했습니다.";
 
 
-    static void printResultMessage() {
+    public static void printResultMessage() {
         System.out.println(EXECUTION_RESULT);
     }
 
-    static void printRacingStatus(Cars cars) {
+    public static void printRacingStatus(Cars cars) {
         printLineBreak();
         cars.getCars().forEach(ResultView::printCarStatus);
     }
 
-    static void printEndGame(RaceWinner raceWinner) {
+    public static void printEndGame(RaceWinner raceWinner) {
         printLineBreak();
         printRacingWinner(raceWinner);
         printRacingEndMessage();
@@ -27,20 +31,20 @@ public class ResultView {
 
     private static void printCarStatus(Car car) {
         printCarName(car);
-        for (int i = 0; i < car.getPosition(); i++) {
+        for (int i = 0; i < car.position(); i++) {
             printCarSymbol(car);
         }
         printLineBreak();
     }
 
     private static void printCarName(Car car) {
-        if (car.getPosition() > 0) {
+        if (car.position() > 0) {
             System.out.print(car.carName() + CAR_NAME_SEPARATOR);
         }
     }
 
     private static void printCarSymbol(Car car) {
-        if (car.getPosition() > 0) {
+        if (car.position() > 0) {
             System.out.print(CAR_SYMBOL);
         }
     }
